Replace flat_router with any_router as unified dispatch interface
Eliminate the two-step router+flat_router workflow by merging the
flattened dispatch directly into any_router. Routes are now inlined
into contiguous arrays at registration time, so dispatch is always
cache-friendly without a separate flattening step.
