+++ src/server/detail/route_match.hpp
@@ -11,7 +11,7 @@
 #define BOOST_HTTP_SERVER_DETAIL_ROUTE_MATCH_HPP
 
 #include <boost/http/detail/config.hpp>
-#include <boost/http/server/basic_router.hpp>
+#include <boost/http/server/any_router.hpp>
 #include "src/server/route_abnf.hpp"
 #include "src/server/detail/stable_string.hpp"
 #include <cstdint>
@@ -20,14 +20,11 @@
 
 namespace boost {
 namespace http {
-namespace detail {
 
 // Matches a path against a pattern
 // Members ordered largest-to-smallest for optimal packing
-struct router_base::matcher
+struct any_router::matcher
 {
-    friend class http::flat_router;
-
     matcher(std::string_view pat, bool end_);
 
     // true if match
@@ -39,29 +36,32 @@ struct router_base::matcher
     system::error_code error() const noexcept { return ec_; }
 
 private:
+    friend class any_router;
+    friend struct any_router::impl;
+
     system::error_code ec_;
     std::string allow_header_;
-    route_pattern pattern_;
+    detail::route_pattern pattern_;
     std::vector<std::string> custom_verbs_;
 
     // 16 bytes (pointer + size)
-    stable_string decoded_pat_;
+    detail::stable_string decoded_pat_;
 
     // 8 bytes each
-    std::size_t first_entry_ = 0;   // flat_router: first entry using this matcher
-    std::size_t skip_ = 0;          // flat_router: entry index to jump to on failure
-    std::uint64_t allowed_methods_ = 0;     // flat_router: bitmask of allowed methods
+    std::size_t first_entry_ = 0;
+    std::size_t skip_ = 0;
+    std::uint64_t allowed_methods_ = 0;
 
     // 4 bytes each
-    opt_flags effective_opts_ = 0;  // flat_router: computed opts for this scope
-    std::uint32_t depth_ = 0;       // flat_router: nesting level (0 = root)
+    opt_flags effective_opts_ = 0;
+    opt_flags own_opts_ = 0;        // router's opt_flags (for re-parenting during inline)
+    std::uint32_t depth_ = 0;
 
     // 1 byte each
     bool end_;      // false for middleware
     bool slash_;
 };
 
-} // detail
 } // http
 } // boost
